Removed the useless TurbineINterface class and commented out code that was giving me errors in the main Articuno class.
For the TurbineMediator class, I added a method that will return an enum based on a turbine Id and a tag. This is used by the main Articuno class to see if a turbine changes state or if wind speed chances, etc. In order to handle it in its queues and lists