Here's the deal.
I have to implement a way for the Turbine to communicate with the ArticunoMain method to update the lists in the main class.

The turbine shouldn't directly talk to the main, so I have to open up a method in the TurbineMediator class as well.

I also updated some tags and tests here and there.