OpcServer:
Commented out the Logging fucntion for quality. Unfortunately, this is being triggered every sampling interval and is
not helpful.

Articuno:
Actually added the turbines in the turbinesWaitingForPause list.

Attempted to shuffle the logCurrentList() call so that not as much stuff gets logged. Unfortunately, I don't think this is working out as I hoped.