Refactoed the MetTower and MetTowerMeidator classes.
Now the Met Tower uses concrete sensor classes and all the other classes can now access the sensors by calling the met tower mediator, then the met tower, then the specific sensor.

Also, the getAllMeasurement is going to be deprecated to an extent. It is not used in any of the main class, but still use in Unit Test for the time being